requirements.txt
Django==2.2.16
mixer==7.1.2
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
Faker==12.0.1
django-debug-toolbar==3.2.4
djangorestframework==3.12.4